Fixed as_poly() by making it raise error with unsupported data types eg: tuples
The commit fixes Issue #18038 in which as_poly() would work with tuples and give wrong output. This was fixed by making it an attribute of Expr as tuples are not a subclass Expr. Tests for the same were added  in sympy/core/tests/test_expr.py and sympy/integrals/tests/test_integrals.py . This ultimately fixed the issue where integrate((x, x)) would give a wrong result.
